refactor : service layer interface 삭제
